requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings>=2.4
pytz==2025.2
pandas
langchain
langchain-openai
httpx
twilio
dateparser
python-multipart
redis>=5.0.0
hiredis>=2.2.0
